+++ Changelog.md
@@ -1,6 +1,10 @@
 ### 3.0.0.rc1 Development
 [Full Changelog](http://github.com/rspec/rspec-rails/compare/v3.0.0.beta2...master)
 
+Enhancements:
+
+* Add `have_http_status` matcher. (Aaron Kromer)
+
 Breaking Changes for 3.0.0:
 
 * Extracts the `mock_model` and `stub_model` methods to the